transform to jsp file
in the html, I just change the type of the button from submit to button.
At first, the onclick of button does not work and I found that is the
type of the button matters. If the type of button is submit, it will
submit the form above and then go into the form action to jump into
another file. Because I am not sure if I can use that action, I change
the type of the button to button. The weekness of this is the website
will not check if the password and user name is empty automatically.
Fortunately, there is function in project to check that but it looks bad
and there is some bugs.